requirements.txt
streamlit
pydantic
pydantic[email]
python-dotenv
psycopg2-binary
sqlalchemy